+++ test/cpp/interop/BUILD
@@ -235,6 +235,32 @@ sh_test(
     tags = ["no_windows"],
 )
 
+grpc_cc_library(
+    name = "xds_stats_watcher",
+    srcs = ["xds_stats_watcher.cc"],
+    hdrs = ["xds_stats_watcher.h"],
+    deps = [
+        ":rpc_behavior_lb_policy",
+        "//:grpc++",
+        "//:grpc++_reflection",
+        "//:grpcpp_admin",
+        "//src/proto/grpc/testing:empty_proto",
+        "//src/proto/grpc/testing:messages_proto",
+        "//src/proto/grpc/testing:test_proto",
+        "//test/core/util:grpc_test_util",
+        "//test/cpp/util:test_config",
+    ],
+)
+
+grpc_cc_test(
+    name = "xds_stats_watcher_test",
+    srcs = [
+        "xds_stats_watcher_test.cc",
+    ],
+    external_deps = ["gtest"],
+    deps = [":xds_stats_watcher"],
+)
+
 grpc_cc_binary(
     name = "xds_interop_client",
     srcs = [
@@ -245,6 +271,7 @@ grpc_cc_binary(
     ],
     deps = [
         ":rpc_behavior_lb_policy",
+        ":xds_stats_watcher",
         "//:grpc++",
         "//:grpc++_reflection",
         "//:grpcpp_admin",
